[EPMEDU-865]: Implement FloatingActionButton for HomeScreen - Implement FloatingActionButton for HomeScreen - Move permissions request and map view from HomeScreenUI file to ui package
