+++ library/resources/src/main/res/drawable-anydpi/ic_geolocation.xml
@@ -1,0 +1,7 @@
+<vector android:height="26dp" android:viewportHeight="24"
+        android:viewportWidth="24" android:width="26dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#000" android:fillType="evenOdd"
+          android:pathData="M12,18C15.314,18 18,15.314 18,12C18,8.686 15.314,6 12,6C8.686,6 6,8.686 6,12C6,15.314 8.686,18 12,18ZM19.938,13C19.487,16.619 16.619,19.487 13,19.938V22C13,22.552 12.552,23 12,23C11.448,23 11,22.552 11,22V19.938C7.381,19.487 4.513,16.619 4.062,13H2C1.448,13 1,12.552 1,12C1,11.448 1.448,11 2,11H4.062C4.513,7.381 7.381,4.513 11,4.062V2C11,1.448 11.448,1 12,1C12.552,1 13,1.448 13,2V4.062C16.619,4.513 19.487,7.381 19.938,11H22C22.552,11 23,11.448 23,12C23,12.552 22.552,13 22,13H19.938Z"/>
+    <path android:fillColor="#000" android:fillType="evenOdd"
+          android:pathData="M12,9.6C10.675,9.6 9.6,10.677 9.6,12.005C9.6,13.334 10.675,14.411 12,14.411C13.326,14.411 14.401,13.334 14.401,12.005C14.401,10.677 13.326,9.6 12,9.6Z"/>
+</vector>
